+++ backend/core/src/seeds/listings/listing-default-seed.ts
@@ -4,6 +4,7 @@ import { BaseEntity, DeepPartial, Repository } from "typeorm"
 import { Listing } from "../../listings/entities/listing.entity"
 import { UnitAccessibilityPriorityType } from "../../unit-accessbility-priority-types/entities/unit-accessibility-priority-type.entity"
 import { UnitType } from "../../unit-types/entities/unit-type.entity"
+import { ReservedCommunityType } from "../../reserved-community-type/entities/reserved-community-type.entity"
 import { AmiChart } from "../../ami-charts/entities/ami-chart.entity"
 import { Property } from "../../property/entities/property.entity"
 import { Unit } from "../../units/entities/unit.entity"
@@ -30,6 +31,8 @@ export class ListingDefaultSeed {
       UnitAccessibilityPriorityType
     >,
     @InjectRepository(UnitType) protected readonly unitTypeRepository: Repository<UnitType>,
+    @InjectRepository(ReservedCommunityType)
+    protected readonly reservedTypeRepository: Repository<ReservedCommunityType>,
     @InjectRepository(AmiChart) protected readonly amiChartRepository: Repository<AmiChart>,
     @InjectRepository(Property) protected readonly propertyRepository: Repository<Property>,
     @InjectRepository(Unit) protected readonly unitsRepository: Repository<Unit>,
@@ -44,7 +47,6 @@ export class ListingDefaultSeed {
     )
     const unitTypeOneBdrm = await this.unitTypeRepository.findOneOrFail({ name: "oneBdrm" })
     const unitTypeTwoBdrm = await this.unitTypeRepository.findOneOrFail({ name: "twoBdrm" })
-
     const amiChart = await this.amiChartRepository.save(getDefaultAmiChart())
 
     const property = await this.propertyRepository.save({