+++ sites/public/src/ListingView.tsx
@@ -12,7 +12,7 @@ import {
   getOccupancyDescription,
   GroupedTable,
   GroupedTableGroup,
-  groupNonReservedAndReservedSummaries,
+  getSummariesTable,
   ImageCard,
   imageUrlFromListing,
   InfoCard,
@@ -33,6 +33,7 @@ import {
   OpenHouseEvent,
   DownloadLotteryResults,
   ReferralApplication,
+  Message,
 } from "@bloom-housing/ui-components"
 import moment from "moment"
 import { ErrorPage } from "../pages/_error"
@@ -82,10 +83,7 @@ export const ListingView = (props: ListingProps) => {
   let groupedUnits: GroupedTableGroup[] = null
 
   if (amiValues.length == 1) {
-    groupedUnits = groupNonReservedAndReservedSummaries(
-      listing.unitsSummarized.byNonReservedUnitType,
-      listing.unitsSummarized.byReservedType
-    )
+    groupedUnits = getSummariesTable(listing.unitsSummarized.byUnitTypeAndRent)
   } // else condition is handled inline below
 
   const occupancyDescription = getOccupancyDescription(listing)
@@ -155,6 +153,15 @@ export const ListingView = (props: ListingProps) => {
     }
   }
 
+  const getReservedTitle = () => {
+    if (
+      listing.reservedCommunityType.name === "senior55" ||
+      listing.reservedCommunityType.name === "senior62"
+    ) {
+      return t("listings.reservedCommunitySeniorTitle")
+    }
+  }
+
   //TODO: Add isReferralApplication boolean field to avoid this logic
   const isReferralApp =
     !listing.applicationDropOffAddress &&
@@ -170,6 +177,11 @@ export const ListingView = (props: ListingProps) => {
         <ImageCard
           title={listing.name}
           imageUrl={imageUrlFromListing(listing, parseInt(process.env.listingPhotoSize))}
+          tagLabel={
+            listing.reservedCommunityType
+              ? t(`listings.reservedCommunityTypes.${props.listing.reservedCommunityType.name}`)
+              : undefined
+          }
         />
         <div className="p-3">
           <p className="font-alt-sans uppercase tracking-widest text-sm font-semibold">
@@ -185,18 +197,20 @@ export const ListingView = (props: ListingProps) => {
       </header>
 
       <div className="w-full md:w-2/3 md:mt-6 md:mb-6 md:px-3 md:pr-8">
+        {listing.reservedCommunityType && (
+          <Message warning={true}>
+            {t("listings.reservedFor", {
+              type: t(`listings.reservedCommunityTypes.${listing.reservedCommunityType.name}`),
+            })}
+          </Message>
+        )}
         {amiValues.length > 1 &&
           amiValues.map((percent) => {
             const byAMI = listing.unitsSummarized.byAMI.find((item) => {
               return parseInt(item.percent, 10) == percent
             })
 
-            groupedUnits = byAMI
-              ? groupNonReservedAndReservedSummaries(
-                  byAMI.byNonReservedUnitType,
-                  byAMI.byReservedType
-                )
-              : []
+            groupedUnits = byAMI ? getSummariesTable(byAMI.byUnitType) : []
 
             return (
               <>
@@ -241,6 +255,19 @@ export const ListingView = (props: ListingProps) => {
           desktopClass="bg-primary-lighter"
         >
           <ul>
+            {listing.reservedCommunityType && (
+              <ListSection title={getReservedTitle()} subtitle={null}>
+                <InfoCard
+                  title={t(`listings.reservedCommunityTypes.${listing.reservedCommunityType.name}`)}
+                  subtitle={t("listings.allUnits")}
+                >
+                  <ExpandableText className="text-sm text-gray-700">
+                    {listing.reservedCommunityDescription}
+                  </ExpandableText>
+                </InfoCard>
+              </ListSection>
+            )}
+
             <ListSection
               title={t("listings.householdMaximumIncome")}
               subtitle={householdMaximumIncomeSubheader}